Add Note to EX188 Certification Preparation
This is a study note for the LinkedIn Learning video:
Red Hat Certified Specialist in Containers (EX188) Certification Preparation
https://www.linkedin.com/learning/red-hat-certified-specialist-in-containers-ex188-cert-prep

This video series covers the usage of Podman, an open-source OCI-
compliant container management tool from Red Hat. It is designed to
handle containers, images, volumes, and pods natively on the Linux
operating system, with support for macOS and MS Windows via a
virtual machine.